Corrected broken file paths
File path for saving config.json was changed from ./saves/config.json to ./config.json in file-management.js.  The reason this was changed was to fix a file path error with the distributed app.

Also included EULA.txt in the package.json extraFiles object so that it packages with the release.  This was so that it would show up in the app Software License Terms menu.