Updated file-management.js
Added functions repairConfig and isConfigValid.  repairConfig makes sure all config properties are valid and adds/resets any that are not.  isConfigValid checks to see if there are any missing properties in the config file.

Also added const defaultConfig, which is the default settings to use when repairing the config file.