Menu Version Number Update
The menu version now automatically matches the version in package.json.  Also added ipcRenderer as a property of the main window, so that remote scripts can communicate with ipcMain on the server side.

Fixed a bug in file-management that caused the config file to unnecessarily repair whenever it was read.